+++ lib/diagnostics/notificationlog.cpp
@@ -1,13 +1,16 @@
 #include "diagnostics/notificationlog.h"
 
+#include "preferences/appsettings.h"
+
 namespace traceview {
 
-NotificationLog::NotificationLog(QObject* parent) : QObject(parent) {}
+NotificationLog::NotificationLog(QObject* parent)
+    : QObject(parent), m_capacity(AppSettings::instance().notificationHistoryCapacity()) {}
 
 void NotificationLog::append(const NotificationEntry& entry) {
     m_entries.append(entry);
     bool droppedOldest = false;
-    if (m_entries.size() > kCapacity) {
+    if (m_entries.size() > m_capacity) {
         m_entries.removeFirst();
         droppedOldest = true;
     }